Version 43.2 for Firefox
Adds the Firefox-specific "applications" key to the manifest, with the
add-on ID that was assigned by Mozilla.  This removes some annoying
validation warnings when uploading new versions.

The minimum Firefox version is set to 48.0, because the chrome APIs that
we're using (chrome.runtime, chrome.extension, chrome.storage) are all
listed as available from "Firefox 48 Android" onward.